requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
click==7.1.1
defusedxml==0.6.0
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
odfpy==1.4.1
pathspec==0.7.0
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
regex==2020.2.20
six==1.14.0
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.1
